% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDV.R
\name{CDV}
\alias{CDV}
\title{Run the analysis}
\usage{
CDV(test.sequences, output.prefix)
}
\arguments{
\item{test.sequences}{is the input fasta or multi-fasta file}

\item{output.prefix}{is a prefix for naming the output files}
}
\description{
This function gets a fasta or multi-fasta file containing full-length
hemmaglutinin nucleotide sequences from CDV and performs the assignment
to the currently described lineages and sublineages, according to their
phylogenetic position.
The function outputs various results including: i) the new alignment, ii) 
the new phylogenetic tree, iii) a table describing the assignments for each
input sequence and iii) a PDF with a schematic representation of the tree.
}

